breaking: Update node version and cleanup.
