+++ packages/config-prettier/package.json
@@ -19,7 +19,7 @@
     "access": "public"
   },
   "engines": {
-    "node": ">=12.17.0",
+    "node": ">=16",
     "npm": ">=6.13.0"
   },
   "files": [
