+++ packages/config-webpack/src/helpers.ts
@@ -1,6 +1,7 @@
 import { Path } from '@beemo/core';
-import { getPackage, WEBPACK_ROOT } from '@oriflame/lumos-common';
+import { getRootPackageJSON } from '@oriflame/lumos-common';
 import ReactRefreshWebpackPlugin from '@pmmmwh/react-refresh-webpack-plugin';
+import glob from 'fast-glob';
 import HtmlWebpackPlugin from 'html-webpack-plugin';
 import path from 'path';
 import webpack, { Configuration, container } from 'webpack';
@@ -9,6 +10,8 @@ import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer';
 import { INVALID_CHARS, NUMBER_REGEX } from './constants';
 import getClientEnvironment from './env';
 import { WebpackOptions } from './types';
+
+export const WEBPACK_ROOT = process.cwd();
 
 export const PROD = process.env.NODE_ENV === 'production';
 export const PORT = 3000;
@@ -97,7 +100,7 @@ export function getPlugins({
 }
 
 export function getUniqueName() {
-  const { name } = getPackage();
+  const { name } = getRootPackageJSON();
 
   return `${name.replace(NUMBER_REGEX, '').replace(INVALID_CHARS, '')}`;
 }
@@ -116,4 +119,66 @@ export function getParallelValue(value: boolean | number | string | undefined): 
   }
 
   return Number(value || 1);
+}
+
+const { WEBPACK_ESM_SCOPES, WEBPACK_ESM_PACKAGES } = process.env;
+
+const esmScopes = ['@ori', '@ori-ui', '@ori-events'];
+const esmPackages: string[] = [];
+
+if (WEBPACK_ESM_SCOPES) {
+  esmScopes.push(...WEBPACK_ESM_SCOPES.split(','));
+}
+
+if (WEBPACK_ESM_PACKAGES) {
+  esmPackages.push(...WEBPACK_ESM_PACKAGES.split(','));
+}
+
+export interface AliasMap {
+  [key: string]: string;
+}
+
+export function getESMAliases(): AliasMap {
+  const aliases: AliasMap = {};
+  const pkg = getRootPackageJSON();
+  const buildTargets = ['lib', 'build', 'dist'];
+
+  glob
+    .sync([`node_modules/{${esmScopes.join(',')}}/*`, `node_modules/{${esmPackages.join(',')}}`], {
+      absolute: true,
+      cwd: WEBPACK_ROOT,
+      onlyDirectories: true,
+      onlyFiles: false,
+    })
+    .forEach((modulePath) => {
+      const packageName = modulePath.split('/node_modules/')[1];
+      const esLessName = packageName.replace(/-es$/, '');
+      const esPath = new Path(modulePath, 'es');
+      const esmPath = new Path(modulePath, 'esm');
+
+      // ori-foo/lib -> ori-foo/esm
+      // optimal/lib -> optimal/esm
+      if (esPath.exists() || esmPath.exists()) {
+        const aliasPath = esPath.exists() ? `${packageName}/es` : `${packageName}/esm`;
+        const aliased = buildTargets.some((targetFolder) => {
+          if (new Path(modulePath, targetFolder).exists()) {
+            aliases[`${packageName}/${targetFolder}`] = aliasPath;
+
+            return true;
+          }
+
+          return false;
+        });
+
+        if (!aliased) {
+          aliases[`${packageName}$`] = aliasPath;
+        }
+
+        // lodash -> lodash-es
+      } else if (packageName.endsWith('-es') && pkg.dependencies && pkg.dependencies[esLessName]) {
+        aliases[esLessName] = packageName;
+      }
+    });
+
+  return aliases;
 }
