+++ packages/eslint-config/README.md
@@ -23,9 +23,6 @@ module.exports = {
     "oriflame/future",
     "oriflame/prettier",
   ],
-  "env": {
-    "node": true,
-  },
 };
 ```
