+++ packages/eslint-config/package.json
@@ -22,7 +22,7 @@
     "check-rules": "ts-node tools/check-rules/index.ts"
   },
   "engines": {
-    "node": ">=12.17.0",
+    "node": ">=16",
     "npm": ">=6.13.0"
   },
   "files": [
