+++ packages/jest-preset/package.json
@@ -26,7 +26,7 @@
     "access": "public"
   },
   "engines": {
-    "node": ">=12.17.0",
+    "node": ">=16",
     "npm": ">=6.13.0"
   },
   "peerDependencies": {
