+++ packages/lumos-common/package.json
@@ -21,14 +21,11 @@
     "access": "public"
   },
   "engines": {
-    "node": ">=12.17.0",
+    "node": ">=16",
     "npm": ">=6.13.0"
   },
   "dependencies": {
-    "@beemo/core": "^2.1.2",
-    "@boost/common": "^3.1.0",
-    "execa": "^5.1.1",
-    "fast-glob": "^3.2.7"
+    "@beemo/core": "^2.1.2"
   },
   "packemon": {
     "format": "lib",
