+++ packages/lumos-common/src/constants.ts
@@ -37,18 +37,11 @@ export const IGNORE_PATHS = [
   'build/',
 ];
 
-export const NODE_VERSION = '12.16';
+export const NODE_VERSION = '14';
 
 export const NODE_TARGET = { node: NODE_VERSION };
 
 export const WEB_TARGET = { browsers: ['last 3 versions', 'not ie > 0'] };
-
-export const SCAFFOLD_DEPS = [
-  '@oriflame/conventional-changelog-lint-config',
-  '@commitlint/cli',
-  'husky',
-  'lint-staged',
-];
 
 export const ROOT = process.env.BEEMO_ROOT || process.cwd();
