+++ packages/lumos-common/src/index.ts
@@ -1,6 +1,3 @@
 export * from './constants';
-export * from './core';
-export * from './git';
-export * from './react';
-export * from './webpack-helpers';
+export * from './types';
 export * from './helpers';
