+++ packages/lumos-common/src/types.ts
@@ -1,0 +1,23 @@
+export interface LumosEnvSetting {
+  targets?:
+    | string[]
+    | string
+    | {
+        browsers?: string[] | string;
+        esmodules?: boolean;
+        node?: string | 'current' | true;
+        safari?: string | 'tp';
+        [key: string]: unknown;
+      };
+  spec?: boolean;
+  loose?: boolean;
+  modules?: 'amd' | 'auto' | 'cjs' | 'commonjs' | 'systemjs' | 'umd' | false;
+  debug?: boolean;
+  include?: Array<RegExp | string>;
+  exclude?: Array<RegExp | string>;
+  useBuiltIns?: 'entry' | 'usage' | false;
+  forceAllTransforms?: boolean;
+  configPath?: string;
+  ignoreBrowserslistConfig?: boolean;
+  shippedProposals?: boolean;
+}
