+++ packages/lumos/package.json
@@ -33,7 +33,7 @@
     "lumos": "./bin/lumos"
   },
   "engines": {
-    "node": ">=12.17.0",
+    "node": ">=16",
     "npm": ">=6.13.0"
   },
   "dependencies": {
