+++ packages/lumos/src/helpers/getSettings.ts
@@ -25,36 +25,134 @@ export interface LumosEnvSetting {
 }
 
 export interface LumosSettings {
+  /**
+   * Build folder depends on library if library=true then default is lib otherwise build
+   * @default 'lib'
+   */
   buildFolder: string;
+  /**
+   * Es modules build folder
+   * @default 'esm'
+   */
   esmBuildFolder: string;
+  /**
+   * Code coverage
+   * @default 75
+   */
   coverage: number;
-  docsFolder: string;
+  /**
+   * Babel env settings
+   * @default {}
+   */
   env: LumosEnvSetting;
+  /**
+   * Enable graphql capabilities
+   * @default false
+   */
   graphql: boolean;
+  /**
+   * Optimizations for library
+   * @default false
+   */
   library: boolean;
+  /**
+   * Enable support for esnext javascript
+   * @default false
+   */
   future: boolean;
+  /**
+   * Enable node optimizations
+   * @default false
+   */
   node: boolean;
+  /**
+   * Enable react support and optimizations
+   * @default false
+   */
   react: boolean;
+  /**
+   * Enable nextjs support
+   */
   nextjs: boolean;
+  /**
+   * Source folder
+   * @default 'src'
+   */
   srcFolder: string;
+  /**
+   * Test folder
+   * @default 'tests'
+   */
   testsFolder: string;
-  declarationDir: string;
+  /**
+   * Types folder
+   * @default 'types'
+   */
   typesFolder: string;
+  /**
+   * Webpack entry point
+   * @default undefined
+   */
   entryPoint?: string;
   publicPath?: string;
+  /**
+   * Change root folder
+   * @default process.cwd()
+   */
   root?: string;
+  /**
+   * Enable declaration only for typescript
+   * @default false
+   */
   declarationOnly?: boolean;
+  /**
+   * Change parallelism
+   * @default 'auto'
+   */
   parallel?: boolean | number | string;
+  /**
+   * Test result file name
+   */
   testResultFileName?: string;
+  /**
+   * Disable babel config generation
+   * @default false
+   */
   emptyBabelConfig: boolean;
+  /**
+   * Allow js in typescript
+   * @default false
+   */
   allowJs: boolean;
+  /**
+   * Skip library check in typescript
+   * @default false
+   */
   skipLibCheck: boolean;
+  /**
+   * Change dev server content base
+   */
   devServerContentBase?: string;
+  /**
+   * Change module federation config
+   */
   moduleFederationConfig?: unknown;
+  /**
+   * Change webpack host
+   */
   host?: string;
-  decorators?: false;
+  /**
+   * Enable shared modules
+   */
   enableSharedModules?: boolean;
+  /**
+   * Path to shared module manifest
+   */
   sharedModulesManifestPath?: string;
+  /**
+   * Enable console mocks
+   * @default true
+   */
   enableConsoleMocks?: boolean;
 }
 
@@ -65,10 +163,9 @@ export function getSettings(tool: Tool): LumosSettings {
   Object.assign(settings, instance.config.settings);
 
   return {
-    buildFolder: 'lib',
+    buildFolder: settings.library ? 'lib' : 'build',
     esmBuildFolder: 'esm',
     coverage: 75,
-    docsFolder: 'docs',
     env: {},
     graphql: false,
     library: false,
@@ -83,8 +180,6 @@ export function getSettings(tool: Tool): LumosSettings {
     allowJs: false,
     skipLibCheck: false,
     enableConsoleMocks: true,
-    decorators: false,
-    declarationDir: 'dts',
     ...settings,
   };
 }
